==> Modulo 3/3-Carga Datos 23-06-2025.ipynb <==

# 1) Introducción al manejo de archivos CSV y configuración inicial

## Temas a analizar:
* Establecer dependencias, importar librerías necesarias
* Verificar compatibilidad, mostrar mensajes si las librerías están instaladas
* Preparar herramientas necesarias
* Documentar

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

print(f"versión instalada de pandas: {pd.__version__}")
print(f"versión instalada de numpy: {np.__version__}")
print(f"versión instalada de matplotlib: {plt.matplotlib.__version__}")


# 2) Leer archivos CSV

## Veremos:
* Creación de Dfs para generar archivos CSV
* Manejo de los diferentes delimitadores
* Opciones adicionales
* Nombres personalizados

#Diccionario de datos generado con datos de prueba para la gestión de archivos CSV
datos_ejemplo = {
    'Nombre': ['Chuck','Steven','Felipe','Nelson'],
    'Edad': [60,70,35,38],
    'Ciudad': ['New York','Seattle','Jaraguá','Santiago'],
    'Sueldo': [300, 400, 500, 600]
}

#Creación del DF a partir del diccionario

df = pd.DataFrame(datos_ejemplo)
df

#Creación de un archivo CSV directo en la raíz del proyecto
#Configuración de INDEX - No inclusión de los indices del df en el CSV
df.to_csv('datos_ejemplo.csv', index=False)

#Creación de un CSV sin encabezado
#Configuración de HEADER - Usa la primera fila como nomre de columnas, por defecto está en 0
#HEADER NONE - No genera fila de encabezado, se puede asignar de forma automática y personalizada
df_sin_header = pd.read_csv('datos_ejemplo.csv', header=None)
print("Extracción correcta de datos desde el CSV pero sin la cabecera de los datos (header)\n")
df_sin_header

#Asignación personalizada de nombres a las columnas
df_nombres_personalizados = pd.read_csv('datos_ejemplo.csv', header=None, names=['Persona', 'Años', 'Ubicación', 'Salario'])
df_nombres_personalizados.head()


#Especificar valores nulos personalizados
datos_con_nulos = {
    'Nombre': ['Juan', 'Marta', '?', 'Ana'],
    'Edad': [60,70,'N/A',38],
    'Ciudad': ['Santiago','Valdivia','Coquimbo','?'],
    'Sueldo': [300, 400, 500, 'N/A']
}

#Crear DF del diccionario
df_nulos = pd.DataFrame(datos_con_nulos)
#Crear CSV a partir del df
df_nulos.to_csv('datos_nulos.csv', index=False)

df_nulos_procesados = pd.read_csv('datos_nulos.csv', na_values=['?','N/A'])
print(df_nulos_procesados.head())
print("Valores nulos encontrados:\n")
print(df_nulos_procesados.isnull().sum())


# 3) Escribir en archivos CSV

## Acciones a realizar:
* Personalizar exportación
* Guardar un df en archivos CSV
* Preparar el Df para la exportación

#Creando nuevo df para la exportación
df_para_exportar = df.copy()

#Exportar de manera bien simple
df_para_exportar.to_csv('datos_para_exportar.csv', index=False)
print('Archivo creado exitosamente')

df_verificacion=pd.read_csv('datos_para_exportar.csv')
print("\n contenido del archivo exportado:")
print(df_verificacion)

#Personalizar la exportación de datos a CSV: Cambiar el delimitador de coma (,) a punto y coma (;)
df_para_exportar.to_csv('4_datos_puntoycoma.csv', index=False, sep =';')
print('Archivo con delimitador ";" creado exitosamente\n')

#Personalizar la exportación de datos a CSV: Cambiar el tipo de codificación
df_para_exportar.to_csv('5_datos_codificacion.csv', index=False, encoding='utf-8')
print('Archivo con codificación personalizada creado exitosamente\n')


df_para_exportar.to_csv('6_datos_sinheaders.csv', index=False, header=False)
print('Archivo con omisión de headers creado exitosamente\n')

#Ejemplo completo de personalización en la exportación de datos a CSV
df_para_exportar.to_csv('7_datos_personalizados.csv', index=False, header=False, sep=';',encoding='utf-8')
print('Archivo personalizado creado exitosamente\n')
print("-Sin indice\n-Delimitador ;\n-Codificación UTF-8\n-Sin Encabezados")

4) Lectura de datos en EXCEL

try:
    import openpyxl
    print("Libreria OpenPyXL está instalada")
except ImportError:
    print("Libreria OpenPyXL NO está instalada")

df_excel_ejemplo = pd.DataFrame({
    'Producto' : ['Laptop','Mouse','Teclado','Monitor'],
    'Precio' : [800,25,60,300],
    'Stock' : [10,20,30,40],
    'Categoría' : ['Tecnología','Accesorios','Accesorios','Tecnología']
})

df_excel_ejemplo.to_excel('8_productos.xlsx',index=False)

df_desde_excel=pd.read_excel('8_productos.xlsx')
print("Datos leidos desde Excel de forma exitosa:\n")
print(df_desde_excel)

#Trabajar con múltiples hojas del archivo EXCEL
with pd.ExcelWriter('9_inventario_multiple.xlsx') as writer:
    #Creación primera HOJA
    df_productos = pd.DataFrame({
        'ID' : [1,2,3],
        'Nombre': ['Laptop','Mouse','Teclado']
    })
    df_productos.to_excel(writer, sheet_name='Productos', index=False)

    #Creación segunda HOJA
    df_ventas = pd.DataFrame({
        'Producto_ID' : [1,2,3],
        'Fecha' : ['20-01-2025','20-01-2025','20-01-2025'],
        'Cantidad' : [6,10,15]
    })
    df_ventas.to_excel(writer, sheet_name='Ventas', index=False)

print("Excel con multiples hojas creado con éxito")

#Leer un Excel, una hoja especifica
df_productos_leidos = pd.read_excel('9_inventario_multiple.xlsx', sheet_name='Productos',index_col=0)
print(f"Datos de hoja productos:\n")
print(df_productos_leidos.head())

#Leer multiples hojas de un Excel
df_dict = pd.read_excel('9_inventario_multiple.xlsx', sheet_name=['Productos', 'Ventas'])
print("Datos del Excel con 2 hojas:\n")
print(df_dict['Productos'])
print("\n",df_dict['Ventas'])

df_todas_hojas = pd.read_excel('9_inventario_multiple.xlsx',sheet_name=None)
print(f"Nombres de todas las hojas:{list(df_todas_hojas.keys())}")

# ACTIVIDAD 1

#Leo el csv y transformo a df
df = pd.read_csv('medallas 18-06-25.csv')
df.head()

#Información general del df
df.info()

#Renombro columnas del df
df.columns = ["Medallas Oro","Medallas Plata","Medallas Bronce","Total Medallas","País"]
df.head()

#Cuento cantidad de nulos por columna
df.isnull().sum()

#Reemplazo nulls por 0
df.fillna(0,inplace=True)
df.head()

#Paso df a un csv nuevo
df.to_csv('Act_medallas23-06-25.csv',index=False)

# ACTIVIDAD 2

df = pd.read_html('https://en.wikipedia.org/wiki/List_of_countries_and_dependencies_by_population')
df=df[0]

df.describe()

df.info()
