==> Modulo 3/6-Pivoteo Pandas.ipynb <==
# 1. Importación de librerías y carga de datos

from itertools import groupby

import pandas as pd
import numpy as np

df = pd.read_csv("datos_ejemplo.csv")
df

# 2.Indexación de datos Jerárquica

#Crear un índice jerarquico con Región y Producto
df_jerarquico = df.set_index(["Región","Producto"]).sort_index()
print("DataFrame con índice jerárquico (Región, Producto):")
df_jerarquico

#Acceder a datos específicos en el df jerarquizado
print("Acceso a datos del 'Centro':")
df_jerarquico.loc["Centro"]

print("Acceso a datos específicos 'Centro' y 'Producto A':")
df_jerarquico.loc[("Centro", "A")]

# 3. Agrupación de datos con groupby

#Agrupación básico por región
print("Ventas totales por región:")
ventas_region = df.groupby("Región")['Ventas'].sum()
ventas_region

print("Estadisticas de unidades por producto:")
estadisticas = df.groupby("Producto")['Unidades'].agg(['mean', 'sum','count','min','max'])
estadisticas

print("Ventas promedio por mes:")
ventas_promedio = df.groupby("Mes")['Ventas'].mean().round(2)
ventas_promedio

#Aplicar multiples funciones en varios campos a la vez a través de 'agg()'
print("Análisis completo por región:")
analisis_region = df.groupby("Región").agg({
    'Ventas' : ['mean', 'sum','count','min','max'],
    'Unidades': ['sum','max','min']
}).round(2)
analisis_region

# 4.Pivoteo de datos

# Crear una tabla pivoteada (dinámica): Regiones vs Productos (suma de ventas)
print("Tabla pivote - Ventas por Región y Producto")
pivot_ventas = df.groupby(['Región','Producto'])['Ventas'].sum().unstack(fill_value=0)
pivot_ventas

print("Tabla pivote - Unidades por Región y Mes")
pivot_unidades = df.groupby(["Mes","Región"])["Unidades"].sum().unstack(fill_value=0)
pivot_unidades

# 5. Despivoteo de datos

#Despivotear tabla 'pivot_ventas'
print("Despivoteo de la tabla de ventas")
despivot_ventas = pivot_ventas.reset_index().melt(
    id_vars = ["Región"],
    var_name = "Producto",
    value_name = "Ventas_total"
)
despivot_ventas[despivot_ventas['Ventas_total']>0]
